+++ K8sUtils/private/Write-PodLog.ps1
@@ -24,9 +24,25 @@ function Write-PodLog {
 
     Write-Header $msg -LogLevel $LogLevel -OutputFile $OutputFile
     kubectl logs --namespace $Namespace $PodName $extraLogParams 2>&1 |
-        Where-Object { $_ -NotMatch 'Error.*: (PodInitializing|ContainerCreating)' } | Tee-Object $OutputFile -Append | Write-Host
-    if ($LASTEXITCODE -ne 0) {
-        Write-Warning "Failed to get logs for pod $PodName ($LASTEXITCODE)"
+        Where-Object { $_ -NotMatch 'Error.*: (PodInitializing|ContainerCreating)' } | Tee-Object $OutputFile -Append | Write-MyHost
+    $getLogsExitCode = $LASTEXITCODE
+    Write-Footer "End logs for $prefix $PodName" -LogLevel $LogLevel -OutputFile $OutputFile
+
+    if ($getLogsExitCode -ne 0) {
+        $msg = "Error getting logs for pod $PodName (exit = $getLogsExitCode), checking status"
+        Write-Header $msg -LogLevel error -OutputFile $OutputFile
+        $state = ,(kubectl get pod $PodName -o jsonpath="{.status.containerStatuses.*.state}" | ConvertFrom-Json -Depth 5)
+        foreach ($s in $state) {
+            # can have running, waiting, or terminated properties
+            if ($s -and (Get-Member -InputObject $s -Name waiting) -and (Get-Member -InputObject $s.waiting -Name reason)) {
+                Write-Status ($s.waiting | Out-String -Width 500) -LogLevel error
+            } elseif ($s -and (Get-Member -InputObject $s -Name terminated) -and (Get-Member -InputObject $s.terminated -Name reason)) {
+                Write-Status ($s.terminated | Out-String -Width 500) -LogLevel error
+            } else {
+                Write-Warning "Didn't get state"
+                Write-Warning ($s | Out-String)
+            }
+        }
+        Write-Footer "End status for $prefix $PodName" -LogLevel error -OutputFile $OutputFile
     }
-    Write-Footer "End logs for $prefix $PodName" -OutputFile $OutputFile
 }