+++ Tools/MinimalDeploy.tests.ps1
@@ -10,7 +10,7 @@ BeforeAll {
 
 Describe "Deploys Minimal API" {
 
-    It "runs without errors" {
+    It "runs hook, init ok" {
         $deploy = Deploy-Minimal -PassThru
 
         Test-Deploy $deploy
@@ -20,7 +20,7 @@ Describe "Deploys Minimal API" {
         Test-MainPod $deploy.PodStatuses[0]
     } -Tag 'Happy'
 
-    It "runs without errors with no init" {
+    It "runs without init ok" {
         $deploy = Deploy-Minimal -PassThru -SkipInit
 
         Test-Deploy $deploy
@@ -30,7 +30,7 @@ Describe "Deploys Minimal API" {
         Test-MainPod $deploy.PodStatuses[0]
     } -Tag 'Happy'
 
-    It "runs without errors with no hook" {
+    It "runs without prehook ok" {
         $deploy = Deploy-Minimal -PassThru -SkipPreHook
 
         Test-Deploy $deploy
@@ -38,96 +38,149 @@ Describe "Deploys Minimal API" {
         Test-MainPod $deploy.PodStatuses[0]
     } -Tag 'Happy'
 
-    It "runs without errors with no hook or init" {
+    It "runs without init or prehook ok" {
         $deploy = Deploy-Minimal -PassThru -SkipPreHook -SkipInit
 
         Test-Deploy $deploy
 
         Test-MainPod $deploy.PodStatuses[0]
+    } -Tag 'Happy',"Shortest"
+
+    It "runs with prehook only ok" {
+        $deploy = Deploy-Minimal -PassThru  -SkipInit -SkipDeploy
+
+        $deploy.Running | Should -Be $false
+        $deploy.ReleaseName | Should -Be 'test'
+        $deploy.PodStatuses | Should -Be $null
+        $deploy.PreHookStatus.Status | Should -Be 'Running'
+        $deploy.RollbackStatus | Should -Be 'DeployedOk'
     } -Tag 'Happy'
 
-    It "does a dry run" {
+    It "runs a dry run" {
         $deploy = Deploy-Minimal -PassThru -DryRun 2>&1 | Out-Null
         $deploy | Should -Be $null
     } -Tag 'Happy'
 
-    It "has the main container crash" {
+    It "has main container crash" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -Fail
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Crash'
-    } -Tag 'Crash','Negative'
+    } -Tag 'Crash','Sad'
 
-    It "has the main container has bad image tag" {
+    It "has main container has bad image tag" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -ImageTag zzz
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'ConfigError' -reason "ErrImageNeverPull"
-    } -Tag 'Config','Negative'
+    } -Tag 'Config','Sad'
 
-    It "has the main container has bad secret name" {
+    It "has the main container with a bad secret name" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -BadSecret
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'ConfigError' -reason "CreateContainerConfigError"
-    } -Tag 'Config','Negative'
+    } -Tag 'Config','Sad'
 
-    It "has the main container that times out" {
-        $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -RunCount 100
-        Test-Deploy $deploy -Running $false
+    It "has the main container too short time out" {
+        $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 3 -RunCount 100
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
+
+        Test-MainPod $deploy.PodStatuses[0] -status 'Timeout' -reason "Possible timeout"
+    } -Tag 'Timeout','Sad'
+
+    It "has the main container time out" {
+        $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 10 -RunCount 100
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
-    } -Tag 'Timeout','Negative'
+    } -Tag 'Timeout','Sad'
 
-    It "gets a temporary startup timeout" {
+    It "has a temporary startup timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 60 -RunCount 10 -StartupProbe
         Test-Deploy $deploy -Running $true
 
         Test-MainPod $deploy.PodStatuses[0]
     } -Tag 'Probe', 'Happy'
 
-    It "gets a temporary startup timeout" {
+    It "has a startup timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 10 -RunCount 10 -StartupProbe
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
-        Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
-    } -Tag 'Negative', 'Timeout', 'Probe'
+        Test-MainPod $deploy.PodStatuses[0] -status 'Timeout' -reason "Possible timeout"
+    } -Tag 'Sad', 'Timeout', 'Probe'
 
     It "has a bad probe" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 120 -Readiness '/fail'
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
         $deploy.PodStatuses[0].LastBadEvents.Count | Should -Be 1
         $deploy.PodStatuses[0].LastBadEvents[0] | Should -BeLike 'Readiness probe failed:*'
-    } -Tag 'Negative','Probe'
+    } -Tag 'Sad','Probe'
 
-    It "has a prehook timeout" {
+    It "has a prehook job top timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -HookRunCount 50 -TimeoutSec 10
-        Test-Deploy $deploy -Running $false -podCount 0
+        Test-Deploy $deploy -Running $false -podCount 0 -RollbackStatus 'RolledBack'
 
         # no pod statuses
-    } -Tag 'Negative', 'Timeout'
+    } -Tag 'Sad', 'Timeout'
+
+    It "has an init failure" {
+        $deploy = Deploy-Minimal -PassThru -SkipPreHook -InitFail
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
+
+        Test-MainPod $deploy.PodStatuses[0] -status 'Crash' -reason "Possible timeout"
+    } -Tag 'Sad', 'Crash'
+
+    It "has init bad config" {
+        $deploy = Deploy-Minimal -PassThru -SkipPreHook -InitTag zzz
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
+
+        Test-MainPod $deploy.PodStatuses[0] -status 'ConfigError' -reason "Possible timeout"
+        $deploy.PodStatuses[0].LastBadEvents.Count | Should -BeGreaterThan 1
+        $deploy.PodStatuses[0].LastBadEvents[1] | Should -Be 'Error: ErrImageNeverPull'
+
+    } -Tag 'Sad', 'Crash'
 
     It "has an init timeout" {
-        $deploy = Deploy-Minimal -PassThru -SkipPreHook -TimeoutSec 10 -InitRunCount 50
-        Test-Deploy $deploy -Running $false
+        $deploy = Deploy-Minimal -PassThru -SkipPreHook -TimeoutSec 5 -InitRunCount 50
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
-        Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
-    } -Tag 'Negative', 'Timeout'
+        Test-MainPod $deploy.PodStatuses[0] -status 'Timeout' -reason "Possible timeout"
+    } -Tag 'Sad', 'Timeout'
 
-    It "has the prehook job's container that times out" {
-        $deploy = Deploy-Minimal -PassThru -SkipInit -HookRunCount 100 -PreHookTimeoutSecs 5
+    It "has prehook job hook timeout" {
+        $deploy = Deploy-Minimal -PassThru -SkipInit -HookRunCount 100 -PreHookTimeoutSecs 5 -Verbose
+        $global:deploy = $deploy
         $deploy.Running | Should -Be $false
+        $deploy.RollbackStatus | Should -Be 'RolledBack'
         $deploy.ReleaseName | Should -Be 'test'
 
-        Test-PreHook $deploy.PreHookStatus # will be running
+        Test-PreHook $deploy.PreHookStatus -status 'Timeout' -containerStatus 'Running'
     } -Tag 'Timeout','Negative'
 
-    It "has prehook crash" {
+    It "has prehook job crash" {
         $deploy = Deploy-Minimal -PassThru -HookFail -TimeoutSecs 20 -PreHookTimeoutSecs 20
-        Test-Deploy $deploy -Running $false -PodCount 0
+        Test-Deploy $deploy -Running $false -PodCount 0 -RollbackStatus 'RolledBack'
+
         $deploy.PreHookStatus.Status | Should -Be 'Crash'
-    } -Tag 'Crash','Negative','New'
+    } -Tag 'Crash','Sad'
+
+    It "has prehook config error" {
+        $deploy = Deploy-Minimal -PassThru -HookTag zzz
+        Test-Deploy $deploy -Running $false -PodCount 0 -RollbackStatus 'RolledBack'
+
+        $deploy.PreHookStatus.Status | Should -Be 'ConfigError'
+        $deploy.PreHookStatus.LastBadEvents.Count | Should -BeGreaterThan 1
+        $deploy.PreHookStatus.LastBadEvents[1] | Should -Be 'Error: ErrImageNeverPull'
+    } -Tag 'Config','Sad'
+
+    It "has prehook timeout" {
+        $deploy = Deploy-Minimal -PassThru -PreHookTimeoutSecs 5 -HookRunCount 100
+        Test-Deploy $deploy -Running $false -PodCount 0 -RollbackStatus 'RolledBack'
+
+        $deploy.PreHookStatus.Status | Should -Be 'Timeout'
+    } -Tag 'Config','Sad'
 }