+++ run.ps1
@@ -20,11 +20,12 @@ param (
     [string[]] $Tasks,
     [switch] $Wait,
     [switch] $DryRun,
-    [string] $K8sUtilsVersion
+    [string] $K8sUtilsVersion,
+    [string] $Repository,
+    [string] $NugetPassword = $env:nuget_password
 )
 
 $currentTask = ""
-$imageName = "minimal"
 
 # execute a script, checking lastexit code
 function executeSB {
@@ -51,10 +52,6 @@ function executeSB {
     }
 }
 
-if ($Tasks -eq "ci") {
-    $Tasks = @('CreateLocalNuget', 'Build', 'Test', 'Pack') # todo sample task expansion
-}
-
 foreach ($currentTask in $Tasks) {
 
     try {
@@ -73,12 +70,19 @@ foreach ($currentTask in $Tasks) {
                 }
             }
             'publishK8sUtils' {
-                if (!$nuget_password) {
-                    throw "nuget_password and K8sUtilsVersion must be set"
+                if (!$NugetPassword -or !$Repository) {
+                    throw "NugetPassword and Repository parameters must be set"
                 }
                 executeSB -RelativeDir "K8sUtils" {
-                    $K8sUtilsVersion
-                    Publish-Module -Repository Loyal -Path . -NuGetApiKey $nuget_password
+                    Publish-Module -Repository $Repository -Path . -NuGetApiKey $NugetPassword
+                }
+            }
+            'test' {
+                executeSB  {
+                    $result = Invoke-Pester -PassThru
+                    Write-Information ($result.tests | Where-Object { $_.executed -and !$_.passed } | Select-Object name, @{n='tags';e={$_.tag -join ','}}, @{n='Error';e={$_.ErrorRecord.DisplayErrorMessage -Replace [Environment]::NewLine,"" }} | Out-String)  -InformationAction Continue
+                    Write-Information "Test results: are in `$test_results" -InformationAction Continue
+                    $global:test_results = $result
                 }
             }
             'upgradeHelm' {
@@ -88,7 +92,7 @@ foreach ($currentTask in $Tasks) {
                     if ($DryRun) {
                         $parms += "--dry-run"
                     }
-                    helm upgrade --install test . -f minimal1_values.yaml --namespace default @parms
+                    helm upgrade --install test . -f minimal_values.yaml --namespace default @parms
                 }
             }
             'uninstallHelm' {